Edited some styles and function call syntax
There was a problem with onclick function call in allPosts section's posts. We were simply passing values inside a quotation to make them string but there was a title where a single quotation existed in a word (eg. Begginer's) . so weh had to escape that single quotation with (\\') so that we can properly pass that whole thing as a string